+++ packages/admin-ui/src/private-apis.ts
@@ -1,0 +1,10 @@
+/**
+ * Internal dependencies
+ */
+import { lock } from './lock-unlock';
+import { UserThemeProvider } from './user-theme-provider';
+
+export const privateApis = {};
+lock( privateApis, {
+	UserThemeProvider,
+} );